Revert "[IA-4620] Invalidate Leo token on sign out"
