+++ src/libs/ajax.js
@@ -7,6 +7,7 @@ import {
   fetchDrsHub,
   fetchGoogleForms,
   fetchMartha,
+  fetchOk,
   fetchOrchestration,
   fetchRawls,
   jsonBody,
@@ -22,7 +23,6 @@ import { Apps } from 'src/libs/ajax/leonardo/Apps';
 import { Disks } from 'src/libs/ajax/leonardo/Disks';
 import { Runtimes } from 'src/libs/ajax/leonardo/Runtimes';
 import { Metrics } from 'src/libs/ajax/Metrics';
-import { fetchOk } from 'src/libs/ajax/network-core/fetch-core';
 import { OAuth2 } from 'src/libs/ajax/OAuth2';
 import { SamResources } from 'src/libs/ajax/SamResources';
 import { Support } from 'src/libs/ajax/Support';
