+++ src/libs/ajax/Metrics.ts
@@ -87,4 +87,3 @@ export const Metrics = (signal?: AbortSignal) => {
     }) as (anonId: string) => Promise<void>,
   };
 };
-export type MetricsContract = ReturnType<typeof Metrics>;
