+++ src/libs/ajax/workflows-app/CbasPact.test.js
@@ -2,8 +2,7 @@ import 'setimmediate';
 
 import { MatchersV3, PactV3, SpecificationVersion } from '@pact-foundation/pact';
 import path from 'path';
-import { fetchFromProxy, jsonBody } from 'src/libs/ajax/ajax-common';
-import { fetchOk } from 'src/libs/ajax/network-core/fetch-core';
+import { fetchFromProxy, fetchOk, jsonBody } from 'src/libs/ajax/ajax-common';
 import { Cbas } from 'src/libs/ajax/workflows-app/Cbas';
 import {
   runSetInputDef,
@@ -15,12 +14,8 @@ import {
 jest.mock('src/libs/ajax/ajax-common', () => ({
   ...jest.requireActual('src/libs/ajax/ajax-common'),
   fetchFromProxy: jest.fn(),
-  authOpts: jest.fn(),
-}));
-
-jest.mock('src/libs/ajax/network-core/fetch-core', () => ({
-  ...jest.requireActual('src/libs/ajax/network-core/fetch-core'),
   fetchOk: jest.fn(),
+  authOpts: jest.fn(),
 }));
 
 jest.mock('src/auth/auth', () => {
